requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pyqrcode
pypng
 